Moved api key attribute, added update to base controller.
