+++ src/Api/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
-﻿using AuthorizationManagement.Api.Models.Internal;
+﻿using AuthorizationManagement.Api.Extensions;
+using AuthorizationManagement.Api.Models.Internal;
 using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -6,7 +7,6 @@ using Microsoft.Azure.Cosmos;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using AuthorizationManagement.Api.Extensions;
 using User = AuthorizationManagement.Api.Models.Internal.User;
 
 
@@ -75,8 +75,7 @@ namespace AuthorizationManagement.Api.Controllers
 
             user = Mapper.Map(userDto, user);
 
-            await Container.ReplaceItemAsync(user, id, new PartitionKey(applicationId), new ItemRequestOptions { IfMatchEtag = user.ETag })
-                .ConfigureAwait(false);
+            await UpdateDocumentAsync(user).ConfigureAwait(false);
 
             return Ok(userDto);
         }
