+++ src/Api/appsettings.json
@@ -7,14 +7,15 @@
     }
   },
   "AllowedHosts": "*",
+  "AuthorizationManagementApiKey": "",
   "KeyVaultName": "",
   "CosmosDb": {
-    "ConnectionStringConfigKeyPath": "", 
+    "ConnectionStringConfigKeyPath": "",
     "ConnectionString": "",
     "DatabaseId": "",
     "ContainerId": "",
     "PreferredRegions": [
     ],
     "ConnectionMode": ""
-  } 
+  }
 }
